requirements.txt
pandas
numpy
torch
torchvision
Pillow
scikit-learn
faiss-cpu
sentence-transformers
flask
requests
opencv-python-headless
matplotlib
seaborn
plotly
tqdm
python-dotenv
webcolors
colorthief
ultralytics
werkzeug
gunicorn
huggingface-hub